Move discard function to Rental controller
