+++ routes/api.php
@@ -7,7 +7,7 @@ use App\Http\Controllers\GenreController;
 use App\Http\Controllers\PasswordResetController;
 use App\Http\Controllers\PolicyController;
 use App\Http\Controllers\UserController;
-use App\Http\Controllers\DiscardBookController;
+use App\Http\Controllers\RentalController;
 use Illuminate\Support\Facades\Route;
 
 Route::post('login', [AuthController::class, 'login']);
@@ -48,7 +48,7 @@ Route::group(['middleware' => 'auth:sanctum'], function () {
         Route::get('/policies' , [PolicyController::class, 'index']);
         Route::patch('policies/{policy}', [PolicyController::class, 'update']);
 
-        Route::post('books/{id}/discard', [DiscardBookController::class, 'discard']);
+        Route::post('books/{id}/discard', [RentalController::class, 'discard']);
 
     });
 });
